Let integration specify astro config location
