+++ astro.config.ts
@@ -3,20 +3,19 @@
 import { defineConfig, type AstroUserConfig } from "astro/config";
 import remarkObsidian from "remark-obsidian";
 import { normalizeFrontmatter } from "./plugins/normalize-fontmatter.remark";
+import localsMarkdown from "./plugins/locals-markdown";
+
 import tailwind from "@astrojs/tailwind";
-import markdownIntegration from "./plugins/renderer";
 
 import cloudflare from "@astrojs/cloudflare";
-
-export const markdown: NonNullable<AstroUserConfig["markdown"]> = {
-  remarkPlugins: [normalizeFrontmatter, remarkObsidian],
-};
 
 // https://astro.build/config
 export default defineConfig({
   output: "hybrid",
-  markdown,
-  integrations: [markdownIntegration(), tailwind()],
+  markdown: {
+    remarkPlugins: [normalizeFrontmatter, remarkObsidian],
+  },
+  integrations: [localsMarkdown({ configUrl: import.meta.url }), tailwind()],
   adapter: cloudflare({
     platformProxy: {
       enabled: true,
