+++ plugins/locals-markdown/middleware.d.ts
@@ -1,0 +1,5 @@
+declare global {
+  const __ASTRO_CONFIG_PATH__: string;
+}
+
+export {};
