+++ src/env.d.ts
@@ -3,7 +3,7 @@
 /// <reference types="@cloudflare/workers-types" />
 
 type Runtime = import("@astrojs/cloudflare").Runtime<Env>;
-type Renderer = import("@plugins/renderer").Renderer;
+type Renderer = import("@plugins/locals-markdown").Renderer;
 
 declare namespace App {
   interface Locals extends Runtime, Renderer {}
